backend/requirements.txt
# NTRIA Backend Requirements
# Nigeria Tax Reform Intelligence Assistant

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Vector Database (JSON-based - no external deps needed)

# LLM & Embeddings
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# Data Processing
pydantic>=2.0.0
pandas>=2.0.0

# PDF Processing
pdfplumber>=0.10.0
pypdf>=3.0.0

# NLP
spacy>=3.7.0
nltk>=3.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# Async
aiohttp>=3.9.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Production
gunicorn>=21.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
